+++ docs/source/release_notes.rst
@@ -6,7 +6,13 @@ Release Notes
 Version 1.0
 -----------
 
-1.0.0-b8 :small:`2020-07-12`
+1.0.0-rc.0 :small:`2020-07-15`
+
+- Fix pickling of :class:`cellrank.tl.Lineage` improvements
+- Add additional options to :func:`cellrank.pl.heatmap`
+- Updated documentation
+
+1.0.0-b.8 :small:`2020-07-12`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 - Add installation options for PETSc and SLEPc
@@ -14,12 +20,12 @@ Version 1.0
 - Add minor :class:`cellrank.tl.Lineage` improvements
 - Fix docstring issues
 
-1.0.0-b2 :small:`2020-07-02`
+1.0.0-b.2 :small:`2020-07-02`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 - Fix installation by including future-fstrings.
 
-1.0.0-b1 :small:`2020-07-02`
+1.0.0-b.1 :small:`2020-07-02`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 - Initial beta pre-release.